Restructure files, create main for real-tests, fix freeing-bug, cleanup tests
