--- abs_tests/02_neg_test.c
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   02_neg_test.c                                      :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 15:59:10 by pvasilan         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../libft/libft.h"
-#include <stdlib.h>
-
-int	null_test(void)
-{
-	if (ft_abs(-5) != abs(-5))
-		return (-1);
-	return (0);
-}
