+++ framework/Makefile
@@ -6,7 +6,7 @@
 #    By: tosuman <timo42@proton.me>                 +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/02/02 19:36:56 by tosuman           #+#    #+#              #
-#    Updated: 2024/02/03 01:13:05 by tosuman          ###   ########.fr        #
+#    Updated: 2024/02/03 17:31:44 by tosuman          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -24,7 +24,7 @@ CC = cc
 RM = /bin/rm -f
 CP = /bin/cp -f
 
-CFLAGS = -Wall -Wextra -Werror -std=c89 -pedantic -Wconversion
+CFLAGS = -O0 -g3 -Wall -Wextra -Werror -std=c89 -pedantic -Wconversion
 
 all: libft $(NAME)
