+++ libft/deque/ddeque_init.c
@@ -6,7 +6,7 @@
 /*   By: tischmid <tischmid@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/22 14:36:56 by tischmid          #+#    #+#             */
-/*   Updated: 2024/02/01 11:58:56 by tosuman          ###   ########.fr       */
+/*   Updated: 2024/02/03 17:53:09 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
