+++ libft/deque/ddeque_print_debug.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ddeque_print.c                                     :+:      :+:    :+:   */
+/*   ddeque_print_debug.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tischmid <tischmid@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: tosuman <timo42@proton.me>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/11/22 14:37:38 by tischmid          #+#    #+#             */
-/*   Updated: 2024/02/03 17:42:48 by tosuman          ###   ########.fr       */
+/*   Created: 2024/02/03 17:50:23 by tosuman           #+#    #+#             */
+/*   Updated: 2024/02/03 17:50:44 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-void	ddeque_print(t_ddeque *ddeque, void (print)(void *data, t_bool first))
+void	ddeque_print_debug(t_ddeque *ddeque)
 {
 	t_ddeque_node	*head;
 	t_ddeque_node	*orig_head;
@@ -21,11 +21,13 @@ void	ddeque_print(t_ddeque *ddeque, void (print)(void *data, t_bool first))
 	orig_head = head;
 	if (!head && ft_printf("Empty ddeque.\n"))
 		return ;
-	print(head->data, TRUE);
+	ft_printf("<\033[31m%p\033[m:\033[32m%p\033[m:\033[33m%p\033[m>",
+		head->prev, head, head->next);
 	while (head->next != orig_head)
 	{
 		head = head->next;
-		print(head->data, FALSE);
+		ft_printf("\n<\033[31m%p\033[m:\033[32m%p\033[m:\033[33m%p\033[m>",
+			head->prev, head, head->next);
 	}
 	ft_printf("\n");
 }
