+++ libft/memory/ft_malloc.c
@@ -6,7 +6,7 @@
 /*   By: tosuman <timo42@proton.me>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/01 07:51:00 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/02 23:25:26 by tosuman          ###   ########.fr       */
+/*   Updated: 2024/02/03 17:49:25 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
