+++ real-tests/abs_tests/04_max_test.c
@@ -1,21 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   03_zero_test.c                                     :+:      :+:    :+:   */
+/*   04_max_test.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 14:35:56 by pvasilan         ###   ########.fr       */
+/*   Updated: 2024/02/03 17:21:13 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../libft/libft.h"
-#include <string.h>
+#include "../../libft/libft.h"
+#include <limits.h>
+#include <stdlib.h>
 
-int	zero_test(void)
+int	max_test(void)
 {
-	if (ft_abs(0) != abs(0))
+	if (ft_abs(INT_MAX) != abs(INT_MAX))
 		return (-1);
 	return (0);
 }
