+++ real-tests/abs_tests/05_min_test.c
@@ -1,24 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   03_bigger_string_test.c                            :+:      :+:    :+:   */
+/*   05_min_test.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tosuman <timo42@proton.me>                 +#+  +:+       +#+        */
+/*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 00:47:54 by tosuman          ###   ########.fr       */
+/*   Updated: 2024/02/03 17:21:16 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../libft/libft.h"
-#include <string.h>
+#include "../../libft/libft.h"
+#include <limits.h>
+#include <stdlib.h>
 
-int	bigger_string_test(void)
+int	min_test(void)
 {
-	const char	*str;
-
-	str = "longstringlongstringlongstringlongstring";
-	if (strlen(str) != ft_strlen(str))
+	if (ft_abs(INT_MIN) != abs(INT_MIN))
 		return (-1);
 	return (0);
 }
