+++ real-tests/abs_tests/abs_launcher.c
@@ -1,20 +1,27 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   min_launcher.c                                     :+:      :+:    :+:   */
+/*   abs_launcher.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/02 21:06:39 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 15:58:47 by pvasilan         ###   ########.fr       */
+/*   Updated: 2024/02/03 17:53:26 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "min_tests.h"
-#include "../framework/libunit.h"
-#include "../libft/libft.h"
+#include "../../framework/libunit.h"
+#include "../../libft/libft.h"
 
-int	min_launcher(void)
+#define ROUTINE_NAME "ABS"
+
+int	pos_test(void);
+int	neg_test(void);
+int	zero_test(void);
+int	max_test(void);
+int	min_test(void);
+
+int	abs_launcher(void)
 {
 	t_ddeque	*tests;
 	int			ret;
@@ -26,10 +33,4 @@ int	min_launcher(void)
 	load_test(tests, "INT_MAX test", max_test);
 	load_test(tests, "INT_MIN test", min_test);
 	return (launch_tests(tests, ROUTINE_NAME));
-}
-
-int	main(void)
-{
-	min_launcher();
-	return (0);
 }
