+++ real-tests/min_tests/01_pos_test.c
@@ -1,24 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   min_tests.h                                        :+:      :+:    :+:   */
+/*   01_pos_test.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/02/02 23:16:35 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 15:58:24 by pvasilan         ###   ########.fr       */
+/*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
+/*   Updated: 2024/02/03 17:21:36 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef ABS_TESTS_H
-# define ABS_TESTS_H
+#include "../../libft/libft.h"
+#include <stdlib.h>
 
-# define ROUTINE_NAME "MIN"
-
-int	pos_test(void);
-int	neg_test(void);
-int	zero_test(void);
-int	max_test(void);
-int	min_test(void);
-
-#endif
+int	pos_test(void)
+{
+	if (ft_abs(5) != abs(5))
+		return (-1);
+	return (0);
+}
