+++ real-tests/min_tests/02_neg_test.c
@@ -1,21 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   01_pos_test.c                                      :+:      :+:    :+:   */
+/*   02_neg_test.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 14:35:11 by pvasilan         ###   ########.fr       */
+/*   Updated: 2024/02/03 17:21:39 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../libft/libft.h"
+#include "../../libft/libft.h"
 #include <stdlib.h>
 
-int	pos_test(void)
+int	neg_test(void)
 {
-	if (ft_abs(5) != abs(5))
+	if (ft_abs(-5) != abs(-5))
 		return (-1);
 	return (0);
 }
