+++ real-tests/min_tests/min_launcher.c
@@ -1,35 +1,36 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   strlen_launcher.c                                  :+:      :+:    :+:   */
+/*   min_launcher.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tosuman <timo42@proton.me>                 +#+  +:+       +#+        */
+/*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/02 21:06:39 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 01:15:09 by tosuman          ###   ########.fr       */
+/*   Updated: 2024/02/03 17:23:43 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "strlen_tests.h"
-#include "../framework/libunit.h"
-#include "../libft/libft.h"
+#include "../../framework/libunit.h"
+#include "../../libft/libft.h"
 
-int	strlen_launcher(void)
+#define ROUTINE_NAME "MIN"
+
+int	pos_test(void);
+int	neg_test(void);
+int	zero_test(void);
+int	max_test(void);
+int	min_test(void);
+
+int	min_launcher(void)
 {
 	t_ddeque	*tests;
 	int			ret;
 
 	tests = ddeque_init();
-	load_test(tests, "Hello test", hello_test);
-	load_test(tests, "NULL test", null_test);
-	load_test(tests, "Bigger string test", bigger_string_test);
+	load_test(tests, "Positive test", pos_test);
+	load_test(tests, "Negative test", neg_test);
 	load_test(tests, "Zero test", zero_test);
-	load_test(tests, "newline test", newline_test);
+	load_test(tests, "INT_MAX test", max_test);
+	load_test(tests, "INT_MIN test", min_test);
 	return (launch_tests(tests, ROUTINE_NAME));
-}
-
-int	main(void)
-{
-	strlen_launcher();
-	return (0);
 }
