+++ real-tests/strlen_tests/01_hello_test.c
@@ -1,22 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   05_min_test.c                                      :+:      :+:    :+:   */
+/*   01_hello_test.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: tosuman <timo42@proton.me>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 14:37:57 by pvasilan         ###   ########.fr       */
+/*   Updated: 2024/02/03 17:20:21 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../libft/libft.h"
-#include <limits.h>
-#include <stdlib.h>
+#include "../../libft/libft.h"
+#include <string.h>
 
-int	min_test(void)
+int	hello_test(void)
 {
-	if (ft_abs(INT_MIN) != abs(INT_MIN))
+	const char	*str;
+
+	str = "hello";
+	if (strlen(str) != ft_strlen(str))
 		return (-1);
 	return (0);
 }
