+++ real-tests/strlen_tests/03_bigger_string_test.c
@@ -1,21 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   01_pos_test.c                                      :+:      :+:    :+:   */
+/*   03_bigger_string_test.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: tosuman <timo42@proton.me>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 14:35:11 by pvasilan         ###   ########.fr       */
+/*   Updated: 2024/02/03 17:20:41 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../libft/libft.h"
-#include <stdlib.h>
+#include "../../libft/libft.h"
+#include <string.h>
 
-int	pos_test(void)
+int	bigger_string_test(void)
 {
-	if (ft_abs(5) != abs(5))
+	const char	*str;
+
+	str = "longstringlongstringlongstringlongstring";
+	if (strlen(str) != ft_strlen(str))
 		return (-1);
 	return (0);
 }
