+++ real-tests/strlen_tests/04_nul_test.c
@@ -1,21 +1,24 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   02_neg_test.c                                      :+:      :+:    :+:   */
+/*   04_nul_test.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pvasilan <pvasilan@student.42berlin.de>    +#+  +:+       +#+        */
+/*   By: tosuman <timo42@proton.me>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/03 00:32:04 by tosuman           #+#    #+#             */
-/*   Updated: 2024/02/03 14:35:30 by pvasilan         ###   ########.fr       */
+/*   Updated: 2024/02/03 17:25:45 by tosuman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../libft/libft.h"
+#include "../../libft/libft.h"
 #include <string.h>
 
-int	null_test(void)
+int	nul_test(void)
 {
-	if (ft_abs(-5) != abs(-5))
+	const char	*str;
+
+	str = "";
+	if (strlen(str) != ft_strlen(str))
 		return (-1);
 	return (0);
 }
